+++ magda-web-client/src/actions/userManagementActions.ts
@@ -111,3 +111,9 @@ export function signOutError(error: any): FacetAction {
         error: error
     };
 }
+
+export function hasResetDatasetCache() {
+    return {
+        type: actionTypes.HAS_RESET_DATASET_CACHE
+    };
+}